PMM-3514 Split slowlog2clickhouse into two: slowlog-loader and slowlog-faker. Modified in schema [u]int* -> float32. Removed sqlx.
