+++ .gitignore
@@ -12,3 +12,4 @@ slowlog2clickhouse
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 logs/
+bin/*
