+++ Gopkg.toml
@@ -1,38 +1,12 @@
-
-# Gopkg.toml example
-#
-# Refer to https://github.com/golang/dep/blob/master/docs/Gopkg.toml.md
-# for detailed Gopkg.toml documentation.
-#
-# required = ["github.com/user/thing/cmd/thing"]
-# ignored = ["github.com/user/project/pkgX", "bitbucket.org/user/project/pkgA/pkgY"]
-#
-# [[constraint]]
-#   name = "github.com/user/project"
-#   version = "1.0.0"
-#
-# [[constraint]]
-#   name = "github.com/user/project2"
-#   branch = "dev"
-#   source = "github.com/myfork/project2"
-#
-# [[override]]
-#  name = "github.com/x/y"
-#  version = "2.4.0"
-
 [prune]
   go-tests = true
   non-go = true
   unused-packages = true
 
 [[constraint]]
-  name = "github.com/jmoiron/sqlx"
-  version = "1.2.0"
-
-[[constraint]]
   name = "github.com/kshvakov/clickhouse"
   version = "1.3.5"
 
 [[constraint]]
-  branch = "master"
+  branch = "PMM-3514-group-by-digest-user-host-db-server"
   name = "github.com/percona/go-mysql"
