+++ Makefile
@@ -1,10 +1,10 @@
-up:
+env-up:
 	mkdir logs
 	docker-compose up -d
 	docker exec ch-server clickhouse client  --query="CREATE DATABASE IF NOT EXISTS pmm;"
 	cat pmm-queries-schema.sql | docker exec -i ch-server clickhouse client -d pmm --multiline --multiquery
 
-down:
+env-down:
 	docker-compose down --volumes
 	rm -rf logs
 
@@ -14,11 +14,10 @@ ch-client:
 ps-client:
 	docker exec -ti ps-server mysql -uroot -psecret
 
-run: build
-	./slowlog2clickhouse
-
 build:
-	go build -o slowlog2clickhouse main.go
+	go build -o bin/slowlogfaker cmd/slowlog-faker/main.go
+	go build -o bin/slowlogloader cmd/slowlog-loader/main.go
 
 build-linux:
-	GOOS=linux go build -o slowlog2clickhouse main.go
+	GOOS=linux go build -o bin/slowlogfaker cmd/slowlog-faker/main.go
+	GOOS=linux go build -o bin/slowlogloader cmd/slowlog-loader/main.go
