+++ pmm-queries-schema.sql
@@ -9,10 +9,10 @@ CREATE TABLE IF NOT EXISTS queries (
   `labels.value` Array(String),
   agent_uuid String,
   period_start DateTime,
-  period_length UInt32,
+  period_length Float32,
   example String,
   example_format Enum8('EXAMPLE' = 0, 'DIGEST' = 1),
-  is_truncated UInt32,
+  is_truncated UInt8,
   example_type Enum8(
     'RANDOM' = 0,
     'SLOWEST' = 1,
@@ -20,67 +20,67 @@ CREATE TABLE IF NOT EXISTS queries (
     'WITH_ERROR' = 3
   ),
   example_metrics String,
-  num_query_with_warnings UInt64,
+  num_queries_with_warnings Float32,
   `warnings.code` Array(String),
-  `warnings.count` Array(UInt64),
-  num_query_with_errors UInt64,
+  `warnings.count` Array(Float32),
+  num_query_with_errors Float32,
   `errors.code` Array(String),
-  `errors.count` Array(UInt64),
-  num_queries UInt64,
-  m_query_time_cnt UInt32,
+  `errors.count` Array(Float32),
+  num_queries Float32,
+  m_query_time_cnt Float32,
   m_query_time_sum Float32,
   m_query_time_min Float32,
   m_query_time_max Float32,
   m_query_time_p99 Float32,
-  m_query_time_hg Array(UInt32),
-  m_lock_time_cnt UInt32,
+  m_query_time_hg Array(Float32),
+  m_lock_time_cnt Float32,
   m_lock_time_sum Float32,
   m_lock_time_min Float32,
   m_lock_time_max Float32,
   m_lock_time_p99 Float32,
   m_lock_time_hg Array(Float32),
-  m_rows_sent_cnt UInt64,
-  m_rows_sent_sum UInt64,
-  m_rows_sent_min UInt64,
-  m_rows_sent_max UInt64,
-  m_rows_sent_p99 UInt64,
-  m_rows_sent_hg Array(UInt64),
-  m_rows_examined_cnt UInt64,
-  m_rows_examined_sum UInt64,
-  m_rows_examined_min UInt64,
-  m_rows_examined_max UInt64,
-  m_rows_examined_p99 UInt64,
-  m_rows_examined_hg Array(UInt64),
-  m_rows_affected_cnt UInt64,
-  m_rows_affected_sum UInt64,
-  m_rows_affected_min UInt64,
-  m_rows_affected_max UInt64,
-  m_rows_affected_p99 UInt64,
-  m_rows_affected_hg Array(UInt64),
-  m_rows_read_cnt UInt64,
-  m_rows_read_sum UInt64,
-  m_rows_read_min UInt64,
-  m_rows_read_max UInt64,
-  m_rows_read_p99 UInt64,
-  m_rows_read_hg Array(UInt64),
-  m_merge_passes_cnt UInt64,
-  m_merge_passes_sum UInt64,
-  m_merge_passes_min UInt64,
-  m_merge_passes_max UInt64,
-  m_merge_passes_p99 UInt64,
-  m_merge_passes_hg Array(UInt64),
-  m_innodb_io_r_ops_cnt UInt64,
-  m_innodb_io_r_ops_sum UInt64,
-  m_innodb_io_r_ops_min UInt64,
-  m_innodb_io_r_ops_max UInt64,
-  m_innodb_io_r_ops_p99 UInt64,
-  m_innodb_io_r_ops_hg Array(UInt64),
-  m_innodb_io_r_bytes_cnt UInt64,
-  m_innodb_io_r_bytes_sum UInt64,
-  m_innodb_io_r_bytes_min UInt64,
-  m_innodb_io_r_bytes_max UInt64,
-  m_innodb_io_r_bytes_p99 UInt64,
-  m_innodb_io_r_bytes_hg Array(UInt64),
+  m_rows_sent_cnt Float32,
+  m_rows_sent_sum Float32,
+  m_rows_sent_min Float32,
+  m_rows_sent_max Float32,
+  m_rows_sent_p99 Float32,
+  m_rows_sent_hg Array(Float32),
+  m_rows_examined_cnt Float32,
+  m_rows_examined_sum Float32,
+  m_rows_examined_min Float32,
+  m_rows_examined_max Float32,
+  m_rows_examined_p99 Float32,
+  m_rows_examined_hg Array(Float32),
+  m_rows_affected_cnt Float32,
+  m_rows_affected_sum Float32,
+  m_rows_affected_min Float32,
+  m_rows_affected_max Float32,
+  m_rows_affected_p99 Float32,
+  m_rows_affected_hg Array(Float32),
+  m_rows_read_cnt Float32,
+  m_rows_read_sum Float32,
+  m_rows_read_min Float32,
+  m_rows_read_max Float32,
+  m_rows_read_p99 Float32,
+  m_rows_read_hg Array(Float32),
+  m_merge_passes_cnt Float32,
+  m_merge_passes_sum Float32,
+  m_merge_passes_min Float32,
+  m_merge_passes_max Float32,
+  m_merge_passes_p99 Float32,
+  m_merge_passes_hg Array(Float32),
+  m_innodb_io_r_ops_cnt Float32,
+  m_innodb_io_r_ops_sum Float32,
+  m_innodb_io_r_ops_min Float32,
+  m_innodb_io_r_ops_max Float32,
+  m_innodb_io_r_ops_p99 Float32,
+  m_innodb_io_r_ops_hg Array(Float32),
+  m_innodb_io_r_bytes_cnt Float32,
+  m_innodb_io_r_bytes_sum Float32,
+  m_innodb_io_r_bytes_min Float32,
+  m_innodb_io_r_bytes_max Float32,
+  m_innodb_io_r_bytes_p99 Float32,
+  m_innodb_io_r_bytes_hg Array(Float32),
   m_innodb_io_r_wait_cnt Float32,
   m_innodb_io_r_wait_sum Float32,
   m_innodb_io_r_wait_min Float32,
@@ -99,61 +99,61 @@ CREATE TABLE IF NOT EXISTS queries (
   m_innodb_queue_wait_max Float32,
   m_innodb_queue_wait_p99 Float32,
   m_innodb_queue_wait_hg Array(Float32),
-  m_innodb_pages_distinct_cnt UInt64,
-  m_innodb_pages_distinct_sum UInt64,
-  m_innodb_pages_distinct_min UInt64,
-  m_innodb_pages_distinct_max UInt64,
-  m_innodb_pages_distinct_p99 UInt64,
-  m_innodb_pages_distinct_hg Array(UInt64),
-  m_query_length_cnt UInt64,
-  m_query_length_sum UInt64,
-  m_query_length_min UInt64,
-  m_query_length_max UInt64,
-  m_query_length_p99 UInt64,
-  m_query_length_hg Array(UInt64),
-  m_bytes_sent_cnt UInt64,
-  m_bytes_sent_sum UInt64,
-  m_bytes_sent_min UInt64,
-  m_bytes_sent_max UInt64,
-  m_bytes_sent_p99 UInt64,
-  m_bytes_sent_hg Array(UInt64),
-  m_tmp_tables_cnt UInt64,
-  m_tmp_tables_sum UInt64,
-  m_tmp_tables_min UInt64,
-  m_tmp_tables_max UInt64,
-  m_tmp_tables_p99 UInt64,
-  m_tmp_tables_hg Array(UInt64),
-  m_tmp_disk_tables_cnt UInt64,
-  m_tmp_disk_tables_sum UInt64,
-  m_tmp_disk_tables_min UInt64,
-  m_tmp_disk_tables_max UInt64,
-  m_tmp_disk_tables_p99 UInt64,
-  m_tmp_disk_tables_hg Array(UInt64),
-  m_tmp_table_sizes_cnt UInt64,
-  m_tmp_table_sizes_sum UInt64,
-  m_tmp_table_sizes_min UInt64,
-  m_tmp_table_sizes_max UInt64,
-  m_tmp_table_sizes_p99 UInt64,
-  m_tmp_table_sizes_hg Array(UInt64),
-  m_qc_hit_sum UInt64,
-  m_full_scan_sum UInt64,
-  m_full_join_sum UInt64,
-  m_tmp_table_sum UInt64,
-  m_tmp_table_on_disk_sum UInt64,
-  m_filesort_sum UInt64,
-  m_filesort_on_disk_sum UInt64,
-  m_select_full_range_join_sum UInt64,
-  m_select_range_sum UInt64,
-  m_select_range_check_sum UInt64,
-  m_sort_range_sum UInt64,
-  m_sort_rows_sum UInt64,
-  m_sort_scan_sum UInt64,
-  m_no_index_used_sum UInt64,
-  m_no_good_index_used_sum UInt64,
+  m_innodb_pages_distinct_cnt Float32,
+  m_innodb_pages_distinct_sum Float32,
+  m_innodb_pages_distinct_min Float32,
+  m_innodb_pages_distinct_max Float32,
+  m_innodb_pages_distinct_p99 Float32,
+  m_innodb_pages_distinct_hg Array(Float32),
+  m_query_length_cnt Float32,
+  m_query_length_sum Float32,
+  m_query_length_min Float32,
+  m_query_length_max Float32,
+  m_query_length_p99 Float32,
+  m_query_length_hg Array(Float32),
+  m_bytes_sent_cnt Float32,
+  m_bytes_sent_sum Float32,
+  m_bytes_sent_min Float32,
+  m_bytes_sent_max Float32,
+  m_bytes_sent_p99 Float32,
+  m_bytes_sent_hg Array(Float32),
+  m_tmp_tables_cnt Float32,
+  m_tmp_tables_sum Float32,
+  m_tmp_tables_min Float32,
+  m_tmp_tables_max Float32,
+  m_tmp_tables_p99 Float32,
+  m_tmp_tables_hg Array(Float32),
+  m_tmp_disk_tables_cnt Float32,
+  m_tmp_disk_tables_sum Float32,
+  m_tmp_disk_tables_min Float32,
+  m_tmp_disk_tables_max Float32,
+  m_tmp_disk_tables_p99 Float32,
+  m_tmp_disk_tables_hg Array(Float32),
+  m_tmp_table_sizes_cnt Float32,
+  m_tmp_table_sizes_sum Float32,
+  m_tmp_table_sizes_min Float32,
+  m_tmp_table_sizes_max Float32,
+  m_tmp_table_sizes_p99 Float32,
+  m_tmp_table_sizes_hg Array(Float32),
+  m_qc_hit_sum Float32,
+  m_full_scan_sum Float32,
+  m_full_join_sum Float32,
+  m_tmp_table_sum Float32,
+  m_tmp_table_on_disk_sum Float32,
+  m_filesort_sum Float32,
+  m_filesort_on_disk_sum Float32,
+  m_select_full_range_join_sum Float32,
+  m_select_range_sum Float32,
+  m_select_range_check_sum Float32,
+  m_sort_range_sum Float32,
+  m_sort_rows_sum Float32,
+  m_sort_scan_sum Float32,
+  m_no_index_used_sum Float32,
+  m_no_good_index_used_sum Float32,
   grpstr String,
-  grpint UInt32,
-  `labint.key` Array(UInt32),
-  `labint.value` Array(UInt32)
+  grpint Float32,
+  `labint.key` Array(Float32),
+  `labint.value` Array(Float32)
 ) ENGINE = MergeTree PARTITION BY toYYYYMM(period_start)
 ORDER BY
   (
